changes done model employee class

+++ EMPManagement/src/main/java/com/employee/repository/EmployeeRepository.java
@@ -5,7 +5,8 @@ import org.springframework.stereotype.Repository;
 
 import com.employee.model.Employee;
 
+
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee,Long>{
+public interface EmployeeRepository extends JpaRepository<Employee,Integer>{
 
 }
